Add flexible reservation usage enforcement spec
This spec proposes a new mechanism to allow operators to define custom
policies that determine whether a given user request to create or update
a lease is valid. A chain of enforcement filters will be called on every
lease request. Two filters will initially be defined: one that checks
that leases do not exceed a length threshold, and one that delegates to
an external HTTP service. Future filter implementations may be included
with Blazar, but there is also a mechanism for operators to provide
their own filter implementations as Python modules.

Change-Id: I376d348a334a09614545b1cadafb3cc76cfa1b1c